kernel: interrupt: create a generic interrupt API
Create a generic interface for drivers to make use of interrupt vectors.
This API should be platform-agnostic enough to allow any driver to make
use of virtually any interrupt vector on any CPU. On x86, the first 32
interrupts are set aside for CPU exceptions, and interrupt 128 is set
aside for system calls.

Signed-off-by: Danny Holman <dholman@gymli.org>